Refactoring some of the names in PropertyMappings to make more sense now that it's not just a property aliasing function.
Adding tests for PropertyModder.

Refactoring the way VersionManagerSession is built for testing to make it a little more flexible, using a fluent api.

Refactoring the way the VersionManagerSession is configured in VersionManager to make it a little more accessible to support tests, then extending with TestVersionManager to give access to that method by itself.