Removing old repository-modification tests...this isn't a use case we care about in reality, so letting it go.
Also, track managed deps from BOMs as actual Dependency instances, so we can compare scopes and exclusions in addition to just versions.

Moving the resolution of effective models up to the top of the BomModder...not sure if this is required, but seemed more predictable.

Other minor changes supporting this.